+++ src/main/java/xyz/tomorrowlearncamp/bookking/domain/common/dto/CustomExceptionDto.java
@@ -1,9 +1,11 @@
-package xyz.tomorrowlearncamp.bookking.domain.common.exception;
+package xyz.tomorrowlearncamp.bookking.domain.common.dto;
 
 import lombok.Getter;
+import lombok.ToString;
 import xyz.tomorrowlearncamp.bookking.domain.common.enums.ErrorMessage;
 
 @Getter
+@ToString()
 public class CustomExceptionDto {
 
 	private final String codeName;
