+++ src/main/java/xyz/tomorrowlearncamp/bookking/domain/common/dto/Response.java
@@ -5,6 +5,9 @@ import com.fasterxml.jackson.annotation.JsonInclude.Include;
 import java.util.List;
 import org.springframework.http.HttpStatus;
 
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
 @JsonInclude(Include.NON_NULL)
 public interface Response<T> {
 
@@ -23,4 +26,28 @@ public interface Response<T> {
 	T getData();
 
 	T getError();
+
+	@Getter
+	@RequiredArgsConstructor
+	class SuccessResponse<T> implements Response<T> {
+		private final T data;
+
+		@Override
+		public T getError() {
+			return null;
+		}
+	}
+
+	@Getter
+	@RequiredArgsConstructor
+	class ErrorResponse<T> implements Response<T> {
+
+		private final HttpStatus status;
+		private final T error;
+
+		@Override
+		public T getData() {
+			return null;
+		}
+	}
 }
